Feat/challenger (#14)
* move basehost and basechild out

* initializer to basehost and basechild

* create message event parser into basehost and basechild

* implement basic challenge function comparing two tx events in deposit and output

* format and lint

* bug fix when loading working tree

* wait one minute if no output exists

* add challenger to cmd

* bug fix initializing height & empty event queue

* use opinit version 1 not 0

* add oracle challenge

* renew challenger

* load pending events & add pending events & latest challenges to status

* add timeout challenge & challenge endpoint & pending event status

* add timeout checker

* show 10 latest challenges as sorted

* lint & format

* Feat/batch (#15)

* change batch spec adding header bytes to batch data

* bump opinit version

* delete oracle option

* bug fix marshal unmarshal batch data

* make broadcaster initialize function to combine prepare func with context

* add logs & num pending events status

* update readme & add challenger status

+++ cmd/opinitd/reset.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	bottypes "github.com/initia-labs/opinit-bots-go/bot/types"
+	bottypes "github.com/initia-labs/opinit-bots/bot/types"
 )
 
 func resetDBCmd(ctx *cmdContext) *cobra.Command {
@@ -18,13 +18,17 @@ func resetDBCmd(ctx *cmdContext) *cobra.Command {
 `,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			botType := bottypes.BotTypeFromString(args[0])
-			switch botType {
-			case bottypes.BotTypeExecutor:
-				dbPath := path.Join(ctx.homePath, string(botType))
-				err := os.RemoveAll(dbPath + ".db")
-				if err != nil {
-					return err
-				}
+			if err := botType.Validate(); err != nil {
+				return err
+			}
+
+			dbPath := path.Join(ctx.homePath, string(botType))
+			err := os.RemoveAll(dbPath + ".db")
+			if err != nil {
+				return err
+			}
+
+			if botType == bottypes.BotTypeExecutor {
 				err = os.RemoveAll(path.Join(ctx.homePath, "batch"))
 				if err != nil {
 					return err
